requirements.txt
gradio==4.44.0
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.0.1
opencv-python==4.8.1.78
scikit-learn==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
mlflow==2.8.1
wandb==0.16.0
optuna==3.4.0
albumentations==1.3.1
tensorboard==2.15.1
plotly==5.17.0
